auth-service/app/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-environ>=0.11
psycopg2-binary>=2.9
gunicorn>=21.0
requests>=2.31
